refactor(CarService): change winner finding logic
