+++ src/main/java/racingcar/repository/CarRepository.java
@@ -19,21 +19,12 @@ public class CarRepository {
 			throw new IllegalArgumentException("[ERROR]: 중복된 차 이름이 있습니다.");
 		}
 	}
-	// TODO: getter 지우기
+
 	public static List<Car> getCars() {
 		return List.copyOf(cars.values());
 	}
 
 	public static void clear() {
 		cars.clear();
-	}
-
-	public static String getPositionToString() {
-		StringBuilder sb = new StringBuilder();
-		cars.values()
-			.stream()
-			.forEach((car) -> sb.append(car.toString()).append("\n"));
-		sb.setLength(sb.length() - 1);
-		return sb.toString();
 	}
 }
